consolidate knockback code
