+++ common/cards/hermits/ijevin-rare.ts
@@ -1,5 +1,4 @@
-import {CardComponent, ObserverComponent, SlotComponent} from '../../components'
-import query from '../../components/query'
+import {CardComponent, ObserverComponent} from '../../components'
 import {GameModel} from '../../models/game-model'
 import {afterAttack} from '../../types/priorities'
 import {hermit} from '../defaults'
@@ -42,34 +41,9 @@ const IJevinRare: Hermit = {
 				if (!attack.isAttacker(component.entity) || attack.type !== 'secondary')
 					return
 
-				const pickCondition = query.every(
-					query.not(query.slot.active),
-					query.not(query.slot.empty),
-					query.slot.opponent,
-					query.slot.hermit,
-					query.slot.canBecomeActive,
-				)
-
-				if (!game.components.exists(SlotComponent, pickCondition)) return
-
-				game.addPickRequest({
-					player: opponentPlayer.entity, // For opponent player to pick
-					id: component.entity,
-					message: 'Choose a new active Hermit from your AFK Hermits.',
-					canPick: pickCondition,
-					onResult(pickedSlot) {
-						if (!pickedSlot.inRow()) return
-						opponentPlayer.changeActiveRow(pickedSlot.row)
-					},
-					onTimeout() {
-						let newActiveRow = game.components.find(
-							SlotComponent,
-							pickCondition,
-						)
-						if (!newActiveRow?.inRow()) return
-						opponentPlayer.changeActiveRow(newActiveRow.row)
-					},
-				})
+				let knockbackPickRequest =
+					opponentPlayer.getKnockbackPickRequest(component)
+				if (knockbackPickRequest) game.addPickRequest(knockbackPickRequest)
 			},
 		)
 	},
