+++ common/cards/hermits/tangotek-rare.ts
@@ -42,13 +42,6 @@ const TangoTekRare: Hermit = {
 				if (!attack.isAttacker(component.entity) || attack.type !== 'secondary')
 					return
 
-				const opponentInactiveRowsPickCondition = query.every(
-					query.slot.opponent,
-					query.slot.hermit,
-					query.not(query.slot.active),
-					query.not(query.slot.empty),
-					query.slot.canBecomeActive,
-				)
 				const playerInactiveRowsPickCondition = query.every(
 					query.slot.currentPlayer,
 					query.slot.hermit,
@@ -63,31 +56,9 @@ const TangoTekRare: Hermit = {
 				])
 
 				// If opponent has hermit they can switch to, add a pick request for them to switch
-				if (
-					game.components.exists(
-						SlotComponent,
-						opponentInactiveRowsPickCondition,
-					)
-				) {
-					game.addPickRequest({
-						player: opponentPlayer.entity,
-						id: component.entity,
-						message: 'Pick a new active Hermit from your afk hermits',
-						canPick: opponentInactiveRowsPickCondition,
-						onResult(pickedSlot) {
-							if (!pickedSlot.inRow()) return
-							opponentPlayer.changeActiveRow(pickedSlot.row)
-						},
-						onTimeout() {
-							let newActiveRow = game.components.find(
-								SlotComponent,
-								opponentInactiveRowsPickCondition,
-							)
-							if (!newActiveRow?.inRow()) return
-							opponentPlayer.changeActiveRow(newActiveRow.row)
-						},
-					})
-				}
+				let knockbackPickRequest =
+					opponentPlayer.getKnockbackPickRequest(component)
+				if (knockbackPickRequest) game.addPickRequest(knockbackPickRequest)
 
 				// If we have an afk hermit, didn't just die, and are not bound in place, add a pick for us to switch
 				if (
