+++ common/cards/single-use/egg.ts
@@ -61,7 +61,7 @@ const Egg: SingleUse = {
 						let afkHermitSlot = pickedSlot
 						if (!afkHermitSlot?.inRow()) return
 
-						opponentPlayer.changeActiveRow(afkHermitSlot.row)
+						opponentPlayer.knockback(afkHermitSlot.row)
 
 						const coinFlip = flipCoin(game, player, component)
 						if (coinFlip[0] === 'heads') {
